Refactored operations validators to have fork versions
 - Previously bls_to_execution_change would have been accepted as an operation before capella

 - had to remove create method to make the structure a little cleaner.
